Check ReadMe for further informations

+++ includes/common_part/common_part_struct.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   common_part_struct.h                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ldutriez <ldutriez@student.42.fr>          +#+  +:+       +#+        */
+/*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/12 14:28:06 by ldutriez          #+#    #+#             */
-/*   Updated: 2021/04/13 17:27:18 by ldutriez         ###   ########.fr       */
+/*   Updated: 2021/04/14 11:12:17 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@
 typedef struct	s_args
 {
 	unsigned int	phi_nb;
-	unsigned int	t_to_die;
-	unsigned int	t_to_eat;
-	unsigned int	t_to_sleep;
+	long			t_to_die;
+	long			t_to_eat;
+	long			t_to_sleep;
 	int				must_eat;
 }				t_args;
